Update to re-compose v0.0.5 and re-implement the template
