+++ android/src/main/java/com/github/whyrising/composetemplate/ui/theme/Colors.kt
@@ -1,0 +1,18 @@
+package com.github.whyrising.composetemplate.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+// TODO: Make your own colors
+val Red200 = Color(0xfff297a2)
+val Red800 = Color(0xffd00036)
+
+val Blue300 = Color(0xff64b5f6)
+val Blue500 = Color(0xff2095f3)
+val Blue700 = Color(0xff1976d2)
+val Blue900 = Color(0xff0d46a1)
+
+val Green300 = Color(0xffaece74)
+val Green600 = Color(0xff7caa2d)
+val Green800 = Color(0xff54831a)
+
+val Grey900 = Color(0xff202124)
