+++ android/src/main/java/com/github/whyrising/composetemplate/ui/theme/Shapes.kt
@@ -1,0 +1,8 @@
+package com.github.whyrising.composetemplate.ui.theme
+
+// TODO: Remove the comment and make your own shapes
+// val TemplateShapes = Shapes(
+//    small = CutCornerShape(topLeft = 8.dp),
+//    medium = CutCornerShape(topLeft = 24.dp),
+//    large = RoundedCornerShape(8.dp)
+// )
