--- android/src/main/java/com/why/template/compose/MyApp.kt
@@ -1,12 +1,0 @@
-package com.why.template.compose
-
-import android.app.Application
-import com.why.template.compose.view.common.init
-
-class MyApp : Application() {
-    override fun onCreate() {
-        super.onCreate()
-
-        init()
-    }
-}
