--- android/src/main/java/com/why/template/compose/data/Route.kt
@@ -1,19 +1,0 @@
-package com.why.template.compose.data
-
-enum class Route {
-    HOME,
-    ABOUT,
-    NONE;
-
-    companion object {
-        fun toRoute(route: String?): Route =
-            when (route?.substringBefore("/")) {
-                HOME.name -> HOME
-                ABOUT.name -> ABOUT
-                null -> HOME
-                else -> throw IllegalArgumentException(
-                    "Route $route is not recognized."
-                )
-            }
-    }
-}
