--- android/src/main/java/com/why/template/compose/data/Spec.kt
@@ -1,8 +1,0 @@
-package com.why.template.compose.data
-
-data class Spec(
-    val topBarTitle: String = "",
-    val counter: Int = 0,
-    val currentPage: Route = Route.NONE,
-    val navigateButtonFlag: Boolean = true
-)
