--- android/src/main/java/com/why/template/compose/view/home/Components.kt
@@ -1,164 +1,0 @@
-package com.why.template.compose.view.home
-
-import android.content.res.Configuration
-import android.os.Build
-import android.util.Log
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Button
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.withStyle
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.github.whyrising.recompose.Keys
-import com.github.whyrising.recompose.dispatch
-import com.github.whyrising.recompose.events.event
-import com.github.whyrising.recompose.regEventDb
-import com.github.whyrising.recompose.regEventFx
-import com.github.whyrising.recompose.regSub
-import com.github.whyrising.recompose.subscribe
-import com.github.whyrising.y.collections.core.get
-import com.github.whyrising.y.collections.core.m
-import com.why.template.compose.R
-import com.why.template.compose.data.Route
-import com.why.template.compose.data.Spec
-import com.why.template.compose.view.common.MyApp
-import com.why.template.compose.view.common.init
-
-@Composable
-fun helloText(name: String): AnnotatedString = buildAnnotatedString {
-    append("Hello ")
-    withStyle(
-        SpanStyle(
-            fontWeight = FontWeight.Bold,
-            color = MaterialTheme.colors.primaryVariant
-        )
-    ) {
-        append(name)
-    }
-    append(" \uD83D\uDE01")
-}
-
-@Composable
-fun Greeting(name: String) {
-    Text(
-        modifier = Modifier.padding(24.dp),
-        text = helloText(name)
-    )
-}
-
-fun regHomePageEvents() {
-    regEventDb(":homePage") { db, vec ->
-        val (_, topBarTitle, currentPage) = vec
-
-        (db as Spec).copy(
-            topBarTitle = topBarTitle as String,
-            currentPage = currentPage as Route,
-            navigateButtonFlag = true
-        )
-    }
-
-    regEventDb(":inc") { db, _ ->
-        (db as Spec).copy(counter = db.counter + 1)
-    }
-
-    regEventFx(":navigate") { cofx, vec ->
-        val viewModel = get(cofx, Keys.db) as Spec
-        val route = vec[1]
-        m(
-            Keys.db to viewModel.copy(navigateButtonFlag = false),
-            Keys.fx to arrayListOf(
-                event(":navigate!", route),
-                event(":print!", "I'm currently in About page. yeeeeeah"),
-                event(Keys.dispatch, event(":inc"))
-            )
-        )
-    }
-
-    regSub<Spec>(":nav-button-flag") { db, _ ->
-        db.navigateButtonFlag
-    }
-}
-
-@Composable
-fun HomePage() {
-    val title = stringResource(R.string.top_bar_home_title)
-
-    regHomePageEvents()
-
-    LaunchedEffect(true) {
-        Log.i("LaunchedEffect", "HomePage")
-        dispatch(event(id = ":homePage", title, Route.HOME, true))
-    }
-
-    Column(
-        modifier = Modifier.fillMaxSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        val apiV = Build.VERSION.SDK_INT
-        Greeting(name = "Android $apiV")
-        Spacer(modifier = Modifier.height(24.dp))
-
-        Text(text = subscribe<String>(arrayListOf(":counter")))
-
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Button(onClick = { dispatch(event(":inc")) }) {
-            Text(text = "Increase")
-        }
-
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Button(
-            onClick = { dispatch(event(":navigate", "${Route.ABOUT}/$apiV")) },
-            enabled = subscribe(event(":nav-button-flag"))
-        ) {
-            Text(text = "Navigate")
-        }
-    }
-}
-
-/*
-*
-* Previews
-*
-* */
-
-@Composable
-private fun HomePageWithTheme() {
-    init()
-
-    MyApp {
-        HomePage()
-    }
-}
-
-@Composable
-@Preview(name = "HomePage Preview - Light Mode")
-private fun HomePagePreview() {
-    HomePageWithTheme()
-}
-
-@Composable
-@Preview(
-    name = "HomePage Preview - Dark Mode",
-    uiMode = Configuration.UI_MODE_NIGHT_YES
-)
-private fun HomePageDarkPreview() {
-    HomePageWithTheme()
-}
