--- android/src/main/java/com/why/template/compose/view/samples/customlayout/intrinsics/Example.kt
@@ -1,57 +1,0 @@
-package com.why.template.compose.view.samples.customlayout.intrinsics
-
-import androidx.compose.foundation.layout.IntrinsicSize
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentWidth
-import androidx.compose.material.Divider
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.why.template.compose.view.theme.MyTheme
-
-@Composable
-fun TwoTexts(modifier: Modifier = Modifier, text1: String, text2: String) {
-    Row(
-        modifier = modifier
-            .padding(horizontal = 8.dp)
-            .height(IntrinsicSize.Min)
-    ) {
-        Text(
-            text = text1,
-            modifier = Modifier
-                .weight(1f)
-                .wrapContentWidth(Alignment.Start)
-        )
-        Divider(
-            color = Color.Black,
-            modifier = Modifier
-                .fillMaxHeight()
-                .width(1.dp)
-        )
-        Text(
-            text = text2,
-            modifier = Modifier
-                .weight(1f)
-                .wrapContentWidth(Alignment.End)
-        )
-    }
-}
-
-@Preview
-@Composable
-fun TwoTextsPreview() {
-    MyTheme {
-        Surface {
-            TwoTexts(text1 = "Hi", text2 = "there")
-        }
-    }
-}
