--- android/src/main/java/com/why/template/compose/view/theme/Shapes.kt
@@ -1,8 +1,0 @@
-package com.why.template.compose.view.theme
-
-// TODO: Remove the comment and make your own shapes
-// val TemplateShapes = Shapes(
-//    small = CutCornerShape(topLeft = 8.dp),
-//    medium = CutCornerShape(topLeft = 24.dp),
-//    large = RoundedCornerShape(8.dp)
-// )
